refactor: change file structure (#67)
