+++ src/components/community/community/AreaCard.tsx
